[ANDROSDK-2115] move itemFilter to own folder
