+++ core/src/main/java/org/hisp/dhis/android/persistence/category/CategoryOptionDB.kt
@@ -11,7 +11,7 @@ import org.hisp.dhis.android.persistence.common.AccessDB
 import org.hisp.dhis.android.persistence.common.BaseNameableObjectDB
 import org.hisp.dhis.android.persistence.common.EntityDB
 import org.hisp.dhis.android.persistence.common.applyBaseNameableFields
-import org.hisp.dhis.android.persistence.common.toDB
+import org.hisp.dhis.android.persistence.itemfilter.toDB
 
 @Entity(
     tableName = "CategoryOption",
