+++ core/src/main/java/org/hisp/dhis/android/persistence/enrollment/EnrollmentDB.kt
@@ -14,7 +14,7 @@ import org.hisp.dhis.android.persistence.common.DeletableObjectDB
 import org.hisp.dhis.android.persistence.common.EntityDB
 import org.hisp.dhis.android.persistence.common.GeometryDB
 import org.hisp.dhis.android.persistence.common.SyncStateDB
-import org.hisp.dhis.android.persistence.common.toDB
+import org.hisp.dhis.android.persistence.itemfilter.toDB
 import org.hisp.dhis.android.persistence.organisationunit.OrganisationUnitDB
 import org.hisp.dhis.android.persistence.program.ProgramDB
 import org.hisp.dhis.android.persistence.trackedentity.TrackedEntityInstanceDB
