+++ core/src/main/java/org/hisp/dhis/android/persistence/event/EventQueryCriteriaDB.kt
@@ -34,7 +34,7 @@ import org.hisp.dhis.android.persistence.common.EntityDB
 import org.hisp.dhis.android.persistence.common.FilterQueryCriteriaDB
 import org.hisp.dhis.android.persistence.common.StringListDB
 import org.hisp.dhis.android.persistence.common.applyFilterQueryCriteriaFields
-import org.hisp.dhis.android.persistence.common.toDB
+import org.hisp.dhis.android.persistence.itemfilter.toDB
 
 internal data class EventQueryCriteriaDB(
     override val followUp: Boolean?,
