+++ core/src/main/java/org/hisp/dhis/android/persistence/note/NoteDB.kt
@@ -10,7 +10,7 @@ import org.hisp.dhis.android.persistence.common.DataObjectDB
 import org.hisp.dhis.android.persistence.common.DeletableObjectDB
 import org.hisp.dhis.android.persistence.common.EntityDB
 import org.hisp.dhis.android.persistence.common.SyncStateDB
-import org.hisp.dhis.android.persistence.common.toDB
+import org.hisp.dhis.android.persistence.itemfilter.toDB
 import org.hisp.dhis.android.persistence.enrollment.EnrollmentDB
 import org.hisp.dhis.android.persistence.event.EventDB
