+++ core/src/main/java/org/hisp/dhis/android/persistence/program/ProgramDB.kt
@@ -20,7 +20,7 @@ import org.hisp.dhis.android.persistence.common.BaseNameableObjectDB
 import org.hisp.dhis.android.persistence.common.EntityDB
 import org.hisp.dhis.android.persistence.common.ObjectWithUidDB
 import org.hisp.dhis.android.persistence.common.applyBaseNameableFields
-import org.hisp.dhis.android.persistence.common.toDB
+import org.hisp.dhis.android.persistence.itemfilter.toDB
 import org.hisp.dhis.android.persistence.trackedentity.TrackedEntityTypeDB
 
 @Entity(
