+++ core/src/main/java/org/hisp/dhis/android/persistence/relationship/RelationshipTypeDB.kt
@@ -10,7 +10,7 @@ import org.hisp.dhis.android.persistence.common.AccessDB
 import org.hisp.dhis.android.persistence.common.BaseIdentifiableObjectDB
 import org.hisp.dhis.android.persistence.common.EntityDB
 import org.hisp.dhis.android.persistence.common.applyBaseIdentifiableFields
-import org.hisp.dhis.android.persistence.common.toDB
+import org.hisp.dhis.android.persistence.itemfilter.toDB
 
 @Entity(
     tableName = "RelationshipType",
