+++ core/src/main/java/org/hisp/dhis/android/persistence/settings/GeneralSettingDB.kt
@@ -8,7 +8,7 @@ import org.hisp.dhis.android.core.util.dateFormat
 import org.hisp.dhis.android.core.util.toJavaDate
 import org.hisp.dhis.android.persistence.common.EntityDB
 import org.hisp.dhis.android.persistence.common.StringListDB
-import org.hisp.dhis.android.persistence.common.toDB
+import org.hisp.dhis.android.persistence.itemfilter.toDB
 
 @Entity(tableName = "GeneralSetting")
 internal data class GeneralSettingDB(
