+++ core/src/main/java/org/hisp/dhis/android/persistence/trackedentity/TrackedEntityInstanceEventFilterDB.kt
@@ -10,7 +10,7 @@ import org.hisp.dhis.android.core.event.EventStatus
 import org.hisp.dhis.android.core.trackedentity.TrackedEntityInstanceEventFilter
 import org.hisp.dhis.android.persistence.common.EntityDB
 import org.hisp.dhis.android.persistence.common.FilterPeriodDB
-import org.hisp.dhis.android.persistence.common.toDB
+import org.hisp.dhis.android.persistence.itemfilter.toDB
 
 @Entity(
     tableName = "TrackedEntityInstanceEventFilter",
